fix: wip - ecr viewer migration provider
