+++ containers/ecr-viewer/src/app/data/db/schemas/common/index.ts
@@ -1,0 +1,3 @@
+export default {
+  "19700101000000_initial.ts": require("./19700101000000_initial"),
+};
