+++ containers/ecr-viewer/src/app/services/instrumentation.ts
@@ -7,8 +7,7 @@ import { Resource } from "@opentelemetry/resources";
 import { NodeSDK, tracing, metrics } from "@opentelemetry/sdk-node";
 import { SEMRESATTRS_SERVICE_NAME } from "@opentelemetry/semantic-conventions";
 
-import { getUnvalidatedDb } from "@/app/api/services/database";
-import { getDbUtils } from "@/app/data/db/utils";
+import { hasPendingMigrations } from "@/app/data/db/utils/migrate";
 
 const sdk = new NodeSDK({
   traceExporter: new OTLPTraceExporter(),
@@ -34,15 +33,12 @@ sdk.start();
 
 // Configuration does not require a database so we don't need to check for pending migrations
 if (!!process.env.METADATA_DATABASE_TYPE) {
-  const dbUtils = getDbUtils();
-  const db = getUnvalidatedDb();
-
-  dbUtils.hasPendingMigrations(db).then((hasPending: boolean) => {
+  hasPendingMigrations().then((hasPending: boolean) => {
     if (hasPending) {
       console.error(
-        "Pending migrations detected. Please run the migration command before starting the server.",
+        // TODO add the route instructions ehre
+        "Pending migrations detected. Please run the migration command.",
       );
-      process.exit(1);
     }
   });
 }
